Issue SC-1727 feat:  Testcase fixes
