+++ data-pipeline-flink/dp-core/src/test/scala/org/sunbird/spec/BaseSpec.scala
@@ -1,0 +1,20 @@
+package org.sunbird.spec
+
+import org.scalatest.{BeforeAndAfterAll, FlatSpec}
+import redis.embedded.RedisServer
+
+class BaseSpec extends FlatSpec with BeforeAndAfterAll{
+  var redisServer: RedisServer = _
+
+  override def beforeAll() {
+    super.beforeAll()
+    redisServer = new RedisServer(6341)
+    redisServer.start()
+  }
+
+  override protected def afterAll(): Unit = {
+    super.afterAll()
+    redisServer.stop()
+  }
+
+}
